Python Microservice/requirements.txt
flask
cmake
face_recognition
matplotlib
numpy
scipy
opencv-python
google-cloud-storage